+++ stripe_terminal/CHANGELOG.md
@@ -1,4 +1,5 @@
 
+## 4.4.0
 - feat: bumped android and ios sdk version to `4.4.0`
 - fix(ios): fixed wrong mapping tap to pay configuration
 - feat(android): configure tap to pay UX. Thanks [@hrueger](https://github.com/hrueger)